requirements.txt
# Core dependencies
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
reportlab>=4.0.0
weasyprint>=60.0

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PDF and HTML generation
jinja2>=3.1.0
markdown>=3.5.0

# Image processing
opencv-python>=4.8.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

